Fix concurrency problem

+++ tests/Master.hx
@@ -24,41 +24,48 @@ class Master {
 			fail('No targets set, use -D targets=php,neko');
 		
 		var result = true;
-		for (container in containers.split(',')) {
-			if (!Context.containers.exists(container))
-				fail('Container $container not available');
-			
-			var server = Context.containers.get(container);
-			try {
-				Sys.println(Ansi.text(Cyan, '\n>> Building container $container'));
-				var process = server(port);
-				waitForConnection(port).handle(function() {
-					for (target in targets.split(',')) {
-						if (!Context.targets.exists(target)) {
-							Ansi.fail('No such target: $target');
-							continue;
-						}
-						Sys.println(Ansi.text(Yellow, '\n>> Running target $target'));
-						var runner = Context.targets.get(target)(port);
-						var code = runner.exitCode();
-						if (code != 0)
-							Ansi.fail('$target failed');
-						result = result && code == 0;
-						if (!result) break;
-					}
-					
-					close(port);
-					process.kill();
-					restoreHxml();
-					Sys.exit(result ? 0 : 1);
-				});
+		var iter = containers.split(',').iterator();
+		
+		function next() {
+			if(iter.hasNext()) {
+				var container = iter.next();
 				
-			} catch(e: Dynamic) {
-				Sys.println(e);
-				Sys.print(CallStack.toString(CallStack.exceptionStack()));
+				if (!Context.containers.exists(container))
+					fail('Container $container not available');
+					
+				var server = Context.containers.get(container);
+
+				try {
+					Sys.println(Ansi.text(Cyan, '\n>> Building container $container'));
+					var process = server(port);
+					waitForConnection(port).handle(function() {
+						for (target in targets.split(',')) {
+							if (!Context.targets.exists(target)) {
+								Ansi.fail('No such target: $target');
+								continue;
+							}
+							Sys.println(Ansi.text(Yellow, '\n>> Running target $target'));
+							var runner = Context.targets.get(target)(port);
+							var code = runner.exitCode();
+							if (code != 0)
+								Ansi.fail('$target failed');
+							result = result && code == 0;
+							if (!result) break;
+						}
+						
+						close(port);
+						process.kill();
+						next();
+					});
+					
+				}
+				
+			} else {
 				restoreHxml();
+				Sys.exit(result ? 0 : 1);
 			}
 		}
+		next();
 	}
 	
 	static function restoreHxml() {
